madres

+++ cfrounds/31/template.cpp
@@ -1,7 +1,7 @@
 #include <bits/stdc++.h>
 using namespace std;
 template <typename T> using minheap = priority_queue<T, vector<T>, greater<T>>;
-using ll = long long;
+using ll = long long int;
 using vi = vector<int>;
 using vlli = vector<ll>;
 using vvi = vector<vi>;
@@ -15,12 +15,13 @@ int msb(long long int x) { return 63 - __builtin_clzll(x);}
 long long int pow2_lb(long long int x) { return (x == (x&-x) ? x : (2 << msb(x)));}
 
 void solve(){
-
+    
 }
+
 
 int main(){
     cin.tie(0)->sync_with_stdio(false);
-    ll t=1;
+    int t=1; 
     cin >> t;
     while(t--) solve();
     return 0;
